# Backend/swagger.js
// Require dependencies
const swaggerAutogen = require('swagger-autogen')
const express = require('swagger-autogen')
const Employee = require('./models/Employee')
const Product = require('./models/Product')

// enable .env file
require('dotenv').config()

// configuration: swagger.json output location and API endpoints
const outputFile = './swagger.json'
const endpointsFiles = ['./index.js']

// Main settings for the swagger/OpenAPI documentation
const swaggerDoc = {
  info: {
    version: "1.0.0",
    title: "Product API for Testing",
    description:"This is a sample API for testing exposes information about Products and Employees. It demonstrates using Express + Mongoose along with API security and swagger/OpenAPI documentation. Documentation has been automatically generated by the swagger-autogen package."
  },
  host:`${process.env.HOSTNAME}:${process.env.PORT}`,
  basePath:"/",
  schemes:["http", "https"],
  consumes:["application/json"],
  produces:["application/json"],

  // Tags are used for grouping endpoints together
  tags:[
    {
      name:"Products",
      description:"Provides CRUD functionality for products."
    },
    {
      name:"Employees",
      description:"Provides CRUD functionality for employees."
    }
  ],
  // Definitions used to define custom object types (schemas)
  definitions:{
    Product: {
      sku:"BR-1234-56",
      name:"Whole meal bread",
       product:[
        {
          $ref:"#/definitions/Product"
        }
      ]
    },
    Employee:{
      empId: 102,
      firstName:"Michael",
      lastName:"Scott",
      username:"mscott",
      password:"DMyc9075@",
      employees:[
        {
          $ref:"#/definitions/Employee"
        }
      ]
    }
  },
  // Defining API security 
  securityDefinitions:{
    Bearer:{
      type:"apiKey",
      in:"header",
      name:"Authorization",  // Name of the HTTP header, query parameter or cookie
      description:"Authorization:Bearer <token>"
    }
  }
}

// Auto-generate the swagger documentation (swagger.json)
swaggerAutogen(outputFile, endpointsFiles, swaggerDoc).then(() => {
  // This runs when the swagger documentation has been generated
  console.log(`The swagger API documentation has been generated to '${outputFile}'`)
}).catch()