# Backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Product API for Testing",
    "description": "This is a sample API for testing exposes information about Products and Employees. It demonstrates using Express + Mongoose along with API security and swagger/OpenAPI documentation. Documentation has been automatically generated by the swagger-autogen package."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Products",
      "description": "Provides CRUD functionality for products."
    },
    {
      "name": "Employees",
      "description": "Provides CRUD functionality for employees."
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Authorization:Bearer <token>"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get all products",
        "description": "Retrieve a listing of all products in the database. If you want a single item, use /api/v1/products{sku} instead.",
        "operationId": "getProducts",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List of all products.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Server-side error.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "some error message..."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create a product",
        "description": "",
        "operationId": "postProduct",
        "parameters": [
          {
            "name": "sku",
            "description": "Code of a product.",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "sku": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "stockOnHand": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/api/v1/products/{sku}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get a product",
        "description": "",
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Error message..."
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Error message..."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Replace a product completely",
        "description": "",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Code of a product."
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "sku": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "stockOnHand": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "Products"
        ],
        "summary": "Update a product completely",
        "description": "",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Code of a product."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete a product",
        "description": "",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Code of a product."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/employees/": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Get list of employees",
        "description": "Retrieving a list of all employees from database.",
        "operationId": "getEmployee",
        "responses": {
          "200": {
            "description": "List of all employees",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "ref": {
                    "type": "string",
                    "example": "#/definitions/Employee"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Server-side error",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Some error message..."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/api/v1/employees/{userName}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Get an employee",
        "description": "",
        "operationId": "getAnEmployee",
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "username",
            "description": "Code of a employee.",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/employees/login": {
      "post": {
        "tags": [
          "Employees"
        ],
        "summary": "Login using username and password.",
        "description": "",
        "operationId": "login",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "definitions": {
    "Product": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "string",
          "example": "BR-1234-56"
        },
        "name": {
          "type": "string",
          "example": "Whole meal bread"
        },
        "product": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        }
      }
    },
    "Employee": {
      "type": "object",
      "properties": {
        "empId": {
          "type": "number",
          "example": 102
        },
        "firstName": {
          "type": "string",
          "example": "Michael"
        },
        "lastName": {
          "type": "string",
          "example": "Scott"
        },
        "username": {
          "type": "string",
          "example": "mscott"
        },
        "password": {
          "type": "string",
          "example": "DMyc9075@"
        },
        "employees": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Employee"
          }
        }
      }
    }
  }
}